Fix codex issue issue:https://gitee.com/openharmony/ace_ace_engine/issues/I53CR5 Signed-off-by: lihucheng <lihucheng1@huawei.com>
 On branch master
 Your branch is up to date with 'origin/master'.

+++ frameworks/bridge/js_frontend/engine/jsi/jsi_base_utils.cpp
@@ -117,8 +117,8 @@ std::string JsiBaseUtils::JsiDumpSourceFile(const std::string& stackStr, const R
     // collect error info first
     for (uint32_t i = 1; i < res.size(); i++) {
         std::string temp = res[i];
-        int32_t closeBracePos = temp.find(closeBrace);
-        int32_t openBracePos = temp.find(openBrace);
+        int32_t closeBracePos = static_cast<int32_t>(temp.find(closeBrace));
+        int32_t openBracePos = static_cast<int32_t>(temp.find(openBrace));
 
         std::string line = "";
         std::string column = "";
